refactor: moved Triton Managers to internal package
