+++ kura/org.eclipse.kura.ai.triton.server/src/main/java/org/eclipse/kura/ai/triton/server/TritonServerServiceOrigImpl.java
@@ -13,6 +13,8 @@
 package org.eclipse.kura.ai.triton.server;
 
 import org.eclipse.kura.ai.triton.server.internal.TritonServerInstanceManager;
+import org.eclipse.kura.ai.triton.server.internal.TritonServerNativeManager;
+import org.eclipse.kura.ai.triton.server.internal.TritonServerRemoteManager;
 import org.eclipse.kura.ai.triton.server.internal.TritonServerServiceAbs;
 import org.eclipse.kura.executor.CommandExecutorService;
