+++ kura/test/org.eclipse.kura.ai.triton.server.test/src/test/java/org/eclipse/kura/ai/triton/server/internal/TritonServerRemoteManagerTest.java
@@ -11,10 +11,11 @@
  *  Eurotech
  ******************************************************************************/
 
-package org.eclipse.kura.ai.triton.server;
+package org.eclipse.kura.ai.triton.server.internal;
 
 import static org.junit.Assert.assertEquals;
 
+import org.eclipse.kura.ai.triton.server.internal.TritonServerRemoteManager;
 import org.junit.Test;
 
 public class TritonServerRemoteManagerTest {
